@@ -1,4 +1,4 @@
-Allincomefoods::Application.routes.draw do
+Rails.application.routes.draw do
   root :to => 'pages#home'
 
   get 'about' => 'pages#about'
